Add color selection menu for node divider
The color picker is part of the board and triggered by clicking the
colored node divider. It applies the selected color as soon as the
mouse is clicked outside. It also keeps track of the used colors for
quick selection and new nodes will automatically get the hue used last.
The lightness is adjusted to the individual themes.

Additional changes:
 - Include icons of Font Awesome from local files