Make color selection part of the menu
The color picker is moved to the menu part acting on all selected nodes
to change multiple at once. Furthermore it can be set before creation
of a new node to initialize it with the configured color. The reset
button will set each selected node to its color before change. Presets
are updated with a change of selection.