Allow for multiple boards managed by the tree-view
The TreeView is augmented to allow for creation and destruction of
custom items in derived classes like the BoardItems in a BoardTree.
All explicit board references are abstracted to object handles.
The HTML structure is accustomed to multiple boards by gathering all
element templates in a template board to get cloned for actual
instances like new boards.

Additional changes:
 - Make color presets span all boards
 - Remove drop-shadow of mini-map
 - Increase font size in menu