Enable saving/loading all boards' contents
The save file content is divided into the containing boards and the
arrangement of those in the tree-view is represented at the top.
Creating a new project will delete all boards and create an empty one,
as there always has to be at least one to work with.

Additional changes:
 - Fix recognition of mouse leave event for tree-view